backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary>=2.9.10
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
tabula-py==2.9.3
openpyxl==3.1.5
jinja2==3.1.4
weasyprint==62.3
slowapi==0.1.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
